LayoutTests/media/video-currentTime-set2-expected.txt

EVENT(load)
TEST(video.currentTime.toFixed(1) == 1.5) OK
TEST(video.currentTime.toFixed(1) == 3.1) OK
END OF TEST
